single-node/requirements.txt
fabric
fabtools
python-digitalocean
